Init folder structure for pages
