+++ src/App.tsx
@@ -1,19 +1,16 @@
 import React from 'react';
-
-import {BrowserRouter, Switch, Route} from 'react-router-dom';
-
-import SignInPage from './SignInPage';
-import ToDoPage from './ToDoPage';
-
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import './App.css';
+import SignInPage from './pages/SignIn';
+import ToDoPage from './pages/ToDo';
 
 function App() {
   return (
     <main className="App">
       <BrowserRouter>
         <Switch>
-          <Route path="/" exact component={SignInPage}/>
-          <Route path="/todo" component={ToDoPage}/>
+          <Route path="/" exact component={SignInPage} />
+          <Route path="/todo" component={ToDoPage} />
         </Switch>
       </BrowserRouter>
     </main>
